+++ docxbox/docx/renderer/docx_renderer.cc
@@ -11,10 +11,9 @@ docx_renderer::Elements docx_renderer::DetectElementType(
 
   std::string identifier = json.substr(2, offset_end_identifier - 2);
 
-  if (identifier == "h1") return Element_Header1;
-  if (identifier == "h2") return Element_Header2;
-  if (identifier == "h3") return Element_Header3;
-  if (identifier == "h4") return Element_Header4;
+  if (identifier == "h1") return Element_Heading1;
+  if (identifier == "h2") return Element_Heading2;
+  if (identifier == "h3") return Element_Heading3;
   if (identifier == "link") return Element_Link;
   if (identifier == "image") return Element_Image;
   if (identifier == "img") return Element_Image;
@@ -26,7 +25,17 @@ docx_renderer::Elements docx_renderer::DetectElementType(
   return Element_None;
 }
 
-bool docx_renderer::IsJsonForImage(const std::string &str) {
+bool docx_renderer::IsValidJsonForHeading(const std::string &str) {
+  Elements kElementType = docx_renderer::DetectElementType(str);
+
+  return helper::String::IsJson(str)
+      && (docx_renderer::Element_Heading1 == kElementType
+          || docx_renderer::Element_Heading2 == kElementType
+          || docx_renderer::Element_Heading3 == kElementType)
+      && helper::String::Contains(str, "text");
+}
+
+bool docx_renderer::IsValidJsonForImage(const std::string &str) {
   return helper::String::IsJson(str)
       && docx_renderer::Element_Image ==
           docx_renderer::DetectElementType(str);