+++ CHANGELOG.md
@@ -3,6 +3,7 @@ docxBox Changelog
 
 V0.0.4 - Not finished yet
 -------------------------
+* Add replace text by heading 1, 2, 3 
 * Add replace text by image 
 * Add cat command: output content of file from within DOCX 
 * Render help in two columns, output also version number